Modularize Entity Renderers
